Add methods to return HAR data in memory.
Added the methods `getHARDataStream` and `getHARData`, returning
a captured HAR file as `BufferedInputStream` or `JSONTokener`,
respectively.

`JSONTokener` was chosen to keep the input stream from being
read until the client user is ready to do so.

Added `FINEST` level logging when entering those messages.

This PR should not be applied until the following are merged:

* https://github.com/saucelabs/saucerest-java/pull/50
* https://github.com/saucelabs/saucerest-java/pull/52