Change `downloadX` methods to return booleans.
By returning Boolean values, it becomes possible to tell if a file was
downloaded successfully or not, based on the return value.

This saves users from having to check if files exist, and provides an easy way
to log that a file downnload failed without having to use the Exception throwing
methods.